requirements.txt
discord.py==2.3.2
groq==0.4.1
flask==3.0.0